refactor: added profiler module
